complete movie search flow with localStorage and UI improvements

+++ src/repository/movieStorage.ts
@@ -1,8 +1,17 @@
 import {SingleMovieInterface} from "../interfaces/SingleMovieInterface";
 import {ApiResponseSuccesInterface} from "../interfaces/ApiResponseSuccesInterface";
+import {SingleMovieSearch} from "../interfaces/movieStorage/SingleMovieSearch";
 
 export function remeberMovieSearch(movie:SingleMovieSearch):void {
     const existingMovie = getAllMoviesSearch()
+    const alreadExists: boolean = existingMovie.some(m =>
+     m.name.toLowerCase() === movie.name.toLowerCase() && m.year === movie.year
+    )
+
+    if(alreadExists){
+        return;
+    }
+
     existingMovie.push(movie);
     localStorage.setItem("remeberedMovies", JSON.stringify(existingMovie));
 
@@ -11,4 +20,6 @@ export function remeberMovieSearch(movie:SingleMovieSearch):void {
 export function getAllMoviesSearch(): SingleMovieSearch[] {
      const data  = localStorage.getItem("remeberedMovies");
      return data ? JSON.parse(data): []
+
+
 }
